@@ -14,13 +14,11 @@
     "babel-preset-react": "^6.1.18",
     "body-parser": "~1.13.2",
     "cookie-parser": "~1.3.5",
-    "css-loader": "^0.16.0",
     "debug": "~2.2.0",
     "ejs": "~2.3.3",
     "eslint-plugin-react": "^3.2.3",
     "express": "~4.13.1",
     "forever": "^0.15.1",
-    "grunt": "^0.4.5",
     "grunt-webpack": "^1.0.8",
     "history": "^1.9.0",
     "html-webpack-plugin": "^1.6.1",
@@ -36,12 +34,11 @@
     "q": "^1.4.1",
     "radium": "^0.14.3",
     "ramda": "^0.17.1",
-    "react-doc-meta": "^0.2.0",
-    "react-geosuggest": "^1.8.0",
-    "react-loader": "^1.4.0",
-    "react-motion": "^0.3.0",
-    "react-pagify": "^0.9.0",
-    "react-router": "^1.0.0-rc1",
+    "react-copy-to-clipboard": "^3.0.4",
+    "react-geosuggest": "^1.13.0",
+    "react-loader": "^2.0.0",
+    "react-motion": "^0.3.1",
+    "react-router": "^1.0.0",
     "rx": "^3.1.0",
     "serve-favicon": "~2.3.0",
     "soundmanager2": "^2.97.20150601-a",
@@ -51,14 +48,10 @@
     "webpack-dev-server": "^1.10.1"
   },
   "devDependencies": {
-    "browserify": "^11.0.1",
-    "flux": "^2.0.3",
     "http-proxy": "^1.11.2",
     "json-loader": "^0.5.3",
-    "react": "^0.13.3",
+    "react": "^0.14.3",
     "react-tools": "^0.13.3",
-    "reactify": "^1.1.1",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.3.1"
+    "vinyl-source-stream": "^1.1.0"
   }
 }
