@@ -3,30 +3,33 @@ import constants from '../constants/constants';
 import HomeSidebar from './HomeSidebar';
 import LoginOverlay from './LoginOverlay';
 
-var HomeView = React.createClass({
+var Home = React.createClass({
 
-	componentDidMount: function() {
+	componentDidMount() {
 		mixpanel.track("User Home Page Open");
 	},
 
-	render: function() {
-		var containerClass = 'flex-row flex-fixed-4x tile-container';
+	showOverLay() {
+		if(loginStatus) {
+			return;
+		} else {
+			return (
+				<LoginOverlay 
+					appState={this.props.appState} 
+					push={this.props.push}
+				/>
+			)
+		}
+	},
 
+	render() {
+		var containerClass = 'flex-row flex-fixed-4x tile-container';
 		var loginStatus = this.props.appState.get('isUserLoggedIn');
 		var user = this.props.appState.get('user');
 
-		if(loginStatus) {
-			var overlay = null;
-		} else {
-			var overlay = (<LoginOverlay 
-					appState={this.props.appState} 
-					push={this.props.push}
-				/>)
-		}
-
 		return (
-			<div id="HomeView" className="view flex-row overlay-container">
-				{overlay}
+			<div id="HomeView" className='flex-row'>
+				{this.showOverlay()}
 				<HomeSidebar 
 					appState={this.props.appState}
 					user={user}
@@ -45,4 +48,4 @@ var HomeView = React.createClass({
 });
 
 
-module.exports = HomeView;
+module.exports = Home;
