@@ -3,14 +3,12 @@ import Loader from 'react-loader';
 import {API_ROOT} from '../constants/constants';
 import SetContainer from './SetContainer';
 
-var TITLE = 'Recent';
-var TYPE = 'set';
-
 var Recent = React.createClass({
 
 	getInitialState() {
 		return {
-			loaded: false
+			loaded: false,
+			sets: []
 		};
 	},
 
@@ -23,43 +21,30 @@ var Recent = React.createClass({
 	},
 
 	getRecentSets() {
-		var push = this.props.push;
-		var recentSets,
-			recentUrl = `${API_ROOT}sets/recent`
-
 		$.ajax({
-			url: recentUrl,
+			url: `${API_ROOT}sets/recent`,
 			type: 'get'
 		})
 		.done(res => {
 			if(res.status === 'success') {
-				recentSets = res.payload.sets_recent;
-				push({
-					type: 'SHALLOW_MERGE',
-					data: {
-						recentBrowseData: recentSets
-					}
-				});
-
 				this.setState({
-					loaded: true
+					loaded: true,
+					sets: res.payload.sets_recent
 				});
 			}
 		});
 	},
 
 	render() {
-		var sets = this.props.appState.get('recentBrowseData');
+		var sets = this.state.sets;
 		var loginStatus = this.props.appState.get('isUserLoggedIn');
 		var user = this.props.appState.get('user');
-		var containerClass = 'flex-row scrollable tile-container';
 		
 		return (
 			<Loader loaded={this.state.loaded}>
 				<SetContainer
 					push={this.props.push}
 					sets={sets}
-					containerClass={containerClass}
 					loginStatus={loginStatus}
 					user={user}
 				/>
