@@ -1,22 +1,13 @@
 import React from 'react';
-import Immutable from 'immutable';
-import Router from 'react-router';
-import DocMeta from 'react-doc-meta';
-import R from 'ramda';
+import {render} from 'react-dom';
+import { Router, IndexRoute, Link, Route, History, Redirect } from 'react-router';
 
-import { IndexRoute, Link, Route, History, Redirect } from 'react-router';
-import GlobalEventHandler from './services/globalEventHandler';
-import {startFacebookSDK} from './services/loginService';
-import detectMobileService from './services/detectMobileService';
-import {API_ROOT} from './constants/constants';
-
+import App from './components/App';
 import Footer from './components/Footer';
-import Header from './components/Header';
-import Player from './components/Player';
 import LandingView from './components/LandingView';
 import EventsView from './components/EventsView';
-import HomeView from './components/HomeView';
-import SetsView from './components/SetsView';
+import Home from './components/Home';
+import Sets from './components/Sets';
 import SearchResultsView from './components/SearchResultsView';
 
 import UpcomingEvents from './components/UpcomingEvents';
@@ -47,186 +38,13 @@ import DMCA from './components/DMCA';
 import Setmusic from './components/Setmusic';
 import SetstoryLandingPage from './components/SetstoryLandingPage';
 import About from './components/About';
-import BlogWrapper from './components/BlogWrapper';
-
-var initialAppState = Immutable.Map({
-	currentSet: {
-		set_length: '00:00',
-		starttime: '00:00',
-		id: null
-	},
-	tracklist: [],
-	currentTrack: null,
-	sound: null,
-	playerHidden: true,
-	playing: false, //change to true once set starts playing
-	timeElapsed: 0, //update while playing
-
-	artistBrowseData: [],
-	festivalBrowseData: [],
-	mixBrowseData: [],
-	activityBrowseData: [],
-	recentBrowseData: [],
-	popularBrowseData: [],
-	featuredEvents: [],
-	upcomingEvents: [],
-	closestEvents: [],
-
-	isUserLoggedIn: false,
-	user: {
-		id: 108
-	},
-	newSets: [],
-	newEvents: [],
-
-	detailId: null,
-	detailData: {
-		"sets": [],
-		"upcomingEvents": [],
-		"links": {
-			"facebook": null,
-			"twitter": null,
-			"instagram": null,
-			"soundcloud": null,
-			"youtube": null
-		}
-	},
-
-	location: {
-		label: 'DEFAULT LOCATION',
-		location: {
-			lat: 29.652175,
-			lng: -82.325856
-		}
-	},
-
-	searchResults: {
-		artists: [],
-		sets: [],
-		upcomingEvents: [],
-		tracks: []
-	}
-});
-
-var evtHandler = GlobalEventHandler(initialAppState);
-var evtTypes = evtHandler.types;
-
-var push = evtHandler.push;
-
-var App = React.createClass({
-	displayName: 'App container',
-	mixins: [History],
-
-	getInitialState() {
-		return {
-			// Let's assume that other ephemeral state
-			// MAY have to exist here.
-			appState: initialAppState
-		};
-	},
-
-	componentWillMount() {
-		this._attachStreams();
-		detectMobileService.detectMobileBrowser();
-		if(!!this.props.params.set) {
-			this.playSet();
-		}
-	},
-
-	componentDidMount() {
-		var metadataPath = window.location.pathname;
-		startFacebookSDK(push);
-
-		// check if user is logged in
-		console.log(this.state.appState.get('isUserLoggedIn'));
-	},
-
-	_attachStreams() {
-		var self = this;
-		evtHandler.floodGate.subscribe(newState => {
-			self.setState({ appState: newState });
-		});
-	},
-
-	playSet() {
-		var id = this.props.params.set;
-		var self = this;
-
-		this.getSetById(id).done(res => {
-			if(res.status === 'success') {
-				var set = res.payload.sets_id;
-				var tracks = set.tracks;
-				var artists = R.pluck('artist', set.artists);
-				var artist = artists.toString().split(',').join(', ');
-
-				var currentSet = {
-					artist: artist,
-					event: set.event.event,
-					id: set.id,
-					set_length: set.set_length,
-					songURL: set.songURL,
-					artistimageURL: set.icon_image.imageURL,
-					starttime: '00:00'
-				};
-
-				push({
-					type: 'SHALLOW_MERGE',
-					data: {
-						currentSet: currentSet,
-						tracklist: tracks,
-						currentTrack: tracks[1].trackname,
-						playing: true
-					}
-				});
-			}
-		});
-	},
-
-	getSetById(id) {
-		return (
-			$.ajax({
-				type: 'get',
-				url: `${API_ROOT}sets/id/${id}`
-			})
-		);
-	},
-
-	render() {
-		var appState = this.state.appState;
-		var tags = [
-			{property: "description", content: "Setmine is a music app dedicated to live events! Relive past music festivals: Ultra, Coachella + more! Find upcoming shows + buy tix + listen to DJs' sets"},
-			{property: "og:site_name", content: "Setmine"},
-			// {property: "og:url", content: "https://setmine.com/metadata/" + encodeURIComponent(metadataPath.substring(1))},
-			{property: "fb:app_id", content: "648288801959503"},
-			{property: "og:description", content: "Setmine offers live music enthusiasts a new way to experience their favorite festival music.  No more struggling to find your favorite sets--we've done it all for you.  Listen to Ultra, Coachella, TomorrowWorld, and many more! Also don't forget to listen your favorite DJ's radio shows!"},
-			{property: "og:image", content: "https://setmine.com/images/setmine-logo-facebook.png"},
-			{property: "og:title", content: "Setmine | View Lineups & Play Sets | Relive Your Favorite Events"},
-			{property: "og:type", content: "website"},
-			{name: "google-site-verification", content: "T4hZD9xTwig_RvyoXaV9XQDYw5ksKEQywRkqaW-CGY4"}
-		];
-		
-		return (
-			<div id='App' className='flex-column'>
-				<DocMeta tags={tags} />
-
-				<Header appState={appState} push={push}/>
-				{
-					React.cloneElement(this.props.children, {
-						appState: appState,
-						push: push
-					})
-				}
-				<Player appState={appState} push={push} />
-			</div>
-		);
-	}
-});
+// import BlogWrapper from './components/BlogWrapper';
 
 var routes = (
 	<Route path='/' component={App}>
 		<IndexRoute component={LandingView}/>
 
-		<Route path='play/:set' component={SetsView}>
+		<Route path='play/:set' component={Sets}>
 			<IndexRoute component={Recent}/>
 			<Route path='mixes' component={Mixes}/>
 			<Route path='popular' component={Popular}/>
@@ -234,7 +52,7 @@ var routes = (
 			<Route path='activities' component={Activities}/>
 		</Route>
 
-		<Route path='sets' component={SetsView}>
+		<Route path='sets' component={Sets}>
 			<IndexRoute component={Recent}/>
 			<Route path='mixes' component={Mixes}/>
 			<Route path='popular' component={Popular}/>
@@ -242,7 +60,7 @@ var routes = (
 			<Route path='activities' component={Activities}/>
 		</Route>
 
-		<Route path='user' component={HomeView}>
+		<Route path='user' component={Home}>
 			<IndexRoute component={Favorites}/>
 			<Route path='sets' component={NewSets}/>
 			<Route path='events' component={NewEvents}/>
@@ -278,15 +96,11 @@ var routes = (
 	</Route>
 );
 
-var headMount = document.getElementById('head-mount-point');
 var bodyMount = document.getElementById('body-mount-point');
 
-import createBrowserHistory from 'history/lib/createBrowserHistory';
-import createHashHistory from 'history/lib/createHashHistory';
-var history = createHashHistory();
-// var history = createBrowserHistory();
+import history from './services/history'
 
-React.render(
+render(
 	<Router history={history}>
 		{routes}
 	</Router>
